requirements_cloud.txt
# Cloud/paper mode requirements — no Playwright, no Streamlit
# Use this on Oracle Cloud / Google Cloud free tier

# Core
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
schedule==1.2.2

# Data providers
yfinance==0.2.40
beautifulsoup4==4.12.3
lxml==5.2.2

# Zerodha (KiteConnect only — no auto-login browser needed for paper mode)
kiteconnect==5.0.1
pyotp==2.9.0

# Notifications
python-telegram-bot==21.4

# Utilities
tabulate==0.9.0
